streamlit_app/requirements.txt
streamlit
requests
python-dotenv
langchain-groq
langchain-huggingface
langchain-community
